Fix twitter button repeat share
