+++ packages/web/src/components/notification/Notification/components/TwitterShareButton.tsx
@@ -1,8 +1,9 @@
-import { useCallback, useEffect, useState } from 'react'
+import { MouseEventHandler, useCallback } from 'react'
 
 import { useDispatch, useSelector } from 'react-redux'
 
 import { ReactComponent as IconTwitterBird } from 'assets/img/iconTwitterBird.svg'
+import { useTwitterButtonStatus } from 'common/hooks/useTwitterButtonStatus'
 import { CommonState } from 'common/store'
 import { fetchUserSocials } from 'common/store/cache/users/actions'
 import { getUser } from 'common/store/cache/users/selectors'
@@ -15,8 +16,6 @@ import styles from './TwitterShareButton.module.css'
 const messages = {
   share: 'Share'
 }
-
-type ShareStatus = 'idle' | 'loading' | 'success'
 
 type StaticTwitterProps = {
   type: 'static'
@@ -41,52 +40,51 @@ export const TwitterShareButton = (props: TwitterShareButtonProps) => {
   const { url = null, ...other } = props
   const record = useRecord()
   const dispatch = useDispatch()
-  const [shareTwitterStatus, setShareTwitterStatus] = useState<ShareStatus>(
-    'idle'
-  )
   const user = useSelector((state: CommonState) =>
     getUser(state, { handle: 'handle' in other ? other.handle : undefined })
   )
-  const userName = user?.name
-  const twitterHandle = user ? user.twitter_handle : null
 
-  const handleClick = useCallback(() => {
-    if (other.type === 'static') {
-      openTwitterLink(url, other.shareText)
-      if (other.analytics) {
-        // @ts-ignore issues with record type
-        record(other.analytics)
+  const {
+    userName,
+    shareTwitterStatus,
+    twitterHandle,
+    setLoading,
+    setIdle
+  } = useTwitterButtonStatus(user)
+
+  const handleClick: MouseEventHandler<HTMLButtonElement> = useCallback(
+    e => {
+      e.stopPropagation()
+      if (other.type === 'static') {
+        openTwitterLink(url, other.shareText)
+        if (other.analytics) {
+          // @ts-ignore issues with record type
+          record(other.analytics)
+        }
       }
-    }
-    if (other.type === 'dynamic') {
-      dispatch(fetchUserSocials(other.handle))
-      setShareTwitterStatus('loading')
-    }
-  }, [url, other, dispatch, record])
-
-  useEffect(() => {
-    if (shareTwitterStatus === 'loading' && twitterHandle !== null) {
-      setShareTwitterStatus('success')
-    }
-  }, [shareTwitterStatus, twitterHandle])
-
-  useEffect(() => {
-    if (
-      other.type === 'dynamic' &&
-      shareTwitterStatus === 'success' &&
-      userName
-    ) {
-      const handle = twitterHandle ? `@${twitterHandle}` : userName
-      const twitterData = other.shareData(handle)
-      if (twitterData) {
-        const { shareText, analytics } = twitterData
-        openTwitterLink(url, shareText)
-        // @ts-ignore issues with record type
-        record(analytics)
-        setShareTwitterStatus('idle')
+      if (other.type === 'dynamic') {
+        dispatch(fetchUserSocials(other.handle))
+        setLoading()
       }
+    },
+    [url, other, dispatch, record, setLoading]
+  )
+
+  if (
+    other.type === 'dynamic' &&
+    shareTwitterStatus === 'success' &&
+    userName
+  ) {
+    const handle = twitterHandle ? `@${twitterHandle}` : userName
+    const twitterData = other.shareData(handle)
+    if (twitterData) {
+      const { shareText, analytics } = twitterData
+      openTwitterLink(url, shareText)
+      // @ts-ignore issues with record type
+      record(analytics)
+      setIdle()
     }
-  }, [other, shareTwitterStatus, twitterHandle, userName, url, record])
+  }
 
   return (
     <button className={styles.root} onClick={handleClick}>
